+++ include/Subdetector.h
@@ -14,7 +14,7 @@ class Subdetector{
     int GetStartLayerBin() const;
     int GetLengthLayerBin() const;				
     int GetNumberOfLayers() const;	
-    int GetLayer(unsigned long const cellID);
+    int GetLayer(unsigned long long const cellID);
     int GetTotCellNumber() const;
 
 	std::vector<float> GetROOTEnergyHisto_binning() const;
